refactor: workflow name
